requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-multipart>=0.0.9

numpy>=1.26
Pillow>=10.3
opencv-contrib-python-headless>=4.9.0   # нужен ximgproc.thinning и LSD
pymupdf>=1.24.4                         # PDF -> image
# для инференса модели и скачивания весов
torch==2.3.1
torchvision==0.18.1
huggingface_hub>=0.23

# геометрия и экспорт в DXF
shapely>=2.0
ezdxf>=1.3

scikit-image>=0.23
shapely>=2.0
